refact: add subscription

+++ internal/storage/storage.go
@@ -51,7 +51,7 @@ type GetSubscriptionsResult struct {
 type Subscription interface {
 	Storage
 	AddSubscription(ctx context.Context, subscription *model.Subscribe) error
-	SubscriptionExist(ctx context.Context, userID int64, sourceID int64) (bool, error)
+	SubscriptionExist(ctx context.Context, userID int64, sourceID uint) (bool, error)
 	GetSubscriptionsByUserID(
 		ctx context.Context, userID int64, opts *GetSubscriptionsOptions,
 	) (*GetSubscriptionsResult, error)
